+++ src/jquery.fancytree.wide.js
@@ -31,7 +31,7 @@
 /**
  * [ext-wide] Recalculate the width of the selection bar after the tree container
  * was resized.<br>
- * Must be called explicitly on window resize, since there is no resize event
+ * May be called explicitly on container resize, since there is no resize event
  * for DIV tags.
  *
  * @alias Fancytree#wideUpdate
@@ -47,6 +47,7 @@ $.ui.fancytree._FancytreeClass.prototype.wideUpdate = function(){
 	//     + 1px title border and 3px title padding
 	inst.lineOfs = (this.options.checkbox ? 3 : 2) * 19;
 	if( prevCw !== inst.contWidth || prevLo !== inst.lineOfs ) {
+		this.debug("wideUpdate: " + inst.contWidth);
 		this.visit(function(node){
 			node.tree._callHook("nodeRenderTitle", node);
 		});
@@ -58,11 +59,12 @@ $.ui.fancytree._FancytreeClass.prototype.wideUpdate = function(){
  */
 $.ui.fancytree.registerExtension({
 	name: "wide",
-	version: "0.0.1",
+	version: "0.0.2",
 	// Default options for this extension.
 	options: {
-		cheap: false,
-		margin: {left: 3, right: 3}
+		autoResize: true,  // call wideUpdate() on window.resize events
+		cheap: false,      // true: use static css only
+		margin: {left: 3, right: 3} // free margins near the selection bar
 	},
 
 	treeCreate: function(ctx){
@@ -76,10 +78,16 @@ $.ui.fancytree.registerExtension({
 		this._local.lineOfs = (ctx.options.checkbox ? 3 : 2) * 19;
 		this._local.levelOfs = 16;
 		this._local.titlePadding = 3;
+		$(window).on("resize" + ctx.widget.eventNamespace, $.ui.fancytree.debounce(200, function(event){
+			if( ctx.options.wide.autoResize && !ctx.options.wide.cheap ) {
+				ctx.tree.wideUpdate();
+			}
+		}));
 	},
-	// treeInit: function(ctx){
-	// 	this._super(ctx);
-	// },
+	treeDestroy: function(ctx){
+		$(window).off("resize" + ctx.widget.eventNamespace);
+		this._super(ctx);
+	},
 	nodeRenderTitle: function(ctx) {
 		var ofs, res, margin,
 			instOpts = ctx.options.wide,